# docs/waterfall-dashboard/openapi-admin.json
{
 "components": {
  "responses": {
...
   "Created": {
    "content": {
     "application/json": {
      "schema": {
       "additionalProperties": true,
       "type": "object"
      }
     }
    },
    "description": "Created"
   },
   "Error": {
    "content": {
     "application/json": {
      "schema": {
       "$ref": "#/components/schemas/Error"
      }
     }
    },
    "description": "Uniform error envelope (doc 04 s1.6)"
   },
   "OK": {
    "content": {
     "application/json": {
      "schema": {
       "additionalProperties": true,
       "type": "object"
      }
     }
    },
    "description": "Success"
   }
  },
  "schemas": {
...
   "Error": {
    "properties": {
     "error": {
      "properties": {
       "code": {
        "type": "string"
       },
       "message": {
        "type": "string"
       }
      },
      "required": [
       "code",
       "message"
      ],
      "type": "object"
     }
    },
    "required": [
     "error"
    ],
    "type": "object"
   },
...
  }
 },
 "info": {
  "description": "Machine-authoritative /v1/admin surface for the Waterfall Management Dashboard (docs/waterfall-dashboard/04-api-contracts.md). Path set is kept at parity with the mounted mux routes by internal/dash/apispec/parity_test.go. Schemas are intentionally shallow; completeness of PATHS is the contract this artifact pins.",
  "title": "Waterfall Dashboard Admin API",
  "version": "1.0.0"
 },
 "openapi": "3.0.3",
 "paths": {
  "/v1/admin/access-log": {
   "get": {
    "operationId": "get_v1_admin_access_log",
    "parameters": [
     {
      "in": "query",
      "name": "user_id",
      "required": false,
      "schema": {
       "type": "string"
      }
     },
     {
      "in": "query",
      "name": "route",
      "required": false,
      "schema": {
       "type": "string"
      }
     },
     {
      "in": "query",
      "name": "status",
      "required": false,
      "schema": {
       "type": "string"
      }
     },
     {
      "in": "query",
      "name": "from",
      "required": false,
      "schema": {
       "type": "string"
      }
     },
     {
      "in": "query",
      "name": "to",
      "required": false,
      "schema": {
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "$ref": "#/components/responses/OK"
     },
     "default": {
      "$ref": "#/components/responses/Error"
     }
    },
    "summary": "API access log (90d)",
    "tags": [
     "security"
    ]
   }
  },
...
  "/v1/admin/audit-log": {
   "get": {
    "operationId": "get_v1_admin_audit_log",
    "parameters": [
     {
      "in": "query",
      "name": "actor_user_id",
      "required": false,
      "schema": {
       "type": "string"
      }
     },
     {
      "in": "query",
      "name": "action",
      "required": false,
      "schema": {
       "type": "string"
      }
     },
     {
      "in": "query",
      "name": "object_kind",
      "required": false,
      "schema": {
       "type": "string"
      }
     },
     {
      "in": "query",
      "name": "object_id",
      "required": false,
      "schema": {
       "type": "string"
      }
     },
     {
      "in": "query",
      "name": "from",
      "required": false,
      "schema": {
       "type": "string"
      }
     },
     {
      "in": "query",
      "name": "to",
      "required": false,
      "schema": {
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "$ref": "#/components/responses/OK"
     },
     "default": {
      "$ref": "#/components/responses/Error"
     }
    },
    "summary": "Hash-chained audit trail",
    "tags": [
     "security"
    ]
   }
  },
  "/v1/admin/audit-log/verify": {
   "get": {
    "operationId": "get_v1_admin_audit_log_verify",
    "parameters": [
     {
      "in": "query",
      "name": "from",
      "required": false,
      "schema": {
       "type": "string"
      }
     },
     {
      "in": "query",
      "name": "to",
      "required": false,
      "schema": {
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "$ref": "#/components/responses/OK"
     },
     "default": {
      "$ref": "#/components/responses/Error"
     }
    },
    "summary": "Walk + verify the audit chain",
    "tags": [
     "security"
    ]
   }
  },
...
  "/v1/admin/auth/login": {
   "post": {
    "operationId": "post_v1_admin_auth_login",
    "responses": {
     "200": {
      "$ref": "#/components/responses/OK"
     },
     "default": {
      "$ref": "#/components/responses/Error"
     }
    },
    "summary": "Verify email+password; start session",
    "tags": [
     "auth"
    ]
   }
  },
  "/v1/admin/auth/logout": {
   "post": {
    "operationId": "post_v1_admin_auth_logout",
    "responses": {
     "200": {
      "$ref": "#/components/responses/OK"
     },
     "default": {
      "$ref": "#/components/responses/Error"
     }
    },
    "summary": "Revoke current session",
    "tags": [
     "auth"
    ]
   }
  },
  "/v1/admin/auth/me": {
   "get": {
    "operationId": "get_v1_admin_auth_me",
    "responses": {
     "200": {
      "$ref": "#/components/responses/OK"
     },
     "default": {
      "$ref": "#/components/responses/Error"
     }
    },
    "summary": "Current user, role, tenant, session expiry",
    "tags": [
     "auth"
    ]
   }
  },
  "/v1/admin/auth/mfa/enroll": {
   "post": {
    "operationId": "post_v1_admin_auth_mfa_enroll",
    "responses": {
     "200": {
      "$ref": "#/components/responses/OK"
     },
     "default": {
      "$ref": "#/components/responses/Error"
     }
    },
    "summary": "Begin TOTP enrollment",
    "tags": [
     "auth"
    ]
   }
  },
  "/v1/admin/auth/mfa/enroll/confirm": {
   "post": {
    "operationId": "post_v1_admin_auth_mfa_enroll_confirm",
    "responses": {
     "200": {
      "$ref": "#/components/responses/OK"
     },
     "default": {
      "$ref": "#/components/responses/Error"
     }
    },
    "summary": "Confirm enrollment; return recovery codes",
    "tags": [
     "auth"
    ]
   }
  },
  "/v1/admin/auth/mfa/verify": {
   "post": {
    "operationId": "post_v1_admin_auth_mfa_verify",
    "responses": {
     "200": {
      "$ref": "#/components/responses/OK"
     },
     "default": {
      "$ref": "#/components/responses/Error"
     }
    },
    "summary": "Complete login with TOTP or recovery code",
    "tags": [
     "auth"
    ]
   }
  },
  "/v1/admin/auth/sessions": {
   "get": {
    "operationId": "get_v1_admin_auth_sessions",
    "parameters": [
     {
      "in": "query",
      "name": "user_id",
      "required": false,
      "schema": {
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "$ref": "#/components/responses/OK"
     },
     "default": {
      "$ref": "#/components/responses/Error"
     }
    },
    "summary": "List sessions",
    "tags": [
     "auth"
    ]
   }
  },
  "/v1/admin/auth/sessions/{id}": {
   "delete": {
    "operationId": "delete_v1_admin_auth_sessions",
    "parameters": [
     {
      "in": "path",
      "name": "id",
      "required": true,
      "schema": {
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "$ref": "#/components/responses/OK"
     },
     "default": {
      "$ref": "#/components/responses/Error"
     }
    },
    "summary": "Revoke a session",
    "tags": [
     "auth"
    ]
   }
  },
...
  "/v1/admin/ip-allowlists": {
   "get": {
    "operationId": "get_v1_admin_ip_allowlists",
    "responses": {
     "200": {
      "$ref": "#/components/responses/OK"
     },
     "default": {
      "$ref": "#/components/responses/Error"
     }
    },
    "summary": "List tenant CIDR allowlist",
    "tags": [
     "users"
    ]
   },
   "put": {
    "operationId": "put_v1_admin_ip_allowlists",
    "responses": {
     "200": {
      "$ref": "#/components/responses/OK"
     },
     "default": {
      "$ref": "#/components/responses/Error"
     }
    },
    "summary": "Replace the allowlist set",
    "tags": [
     "users"
    ]
   }
  },
...
  "/v1/admin/roles": {
   "get": {
    "operationId": "get_v1_admin_roles",
    "responses": {
     "200": {
      "$ref": "#/components/responses/OK"
     },
     "default": {
      "$ref": "#/components/responses/Error"
     }
    },
    "summary": "Static role x action matrix",
    "tags": [
     "users"
    ]
   }
  },
...
  "/v1/admin/settings/mfa-policy": {
   "get": {
    "operationId": "get_v1_admin_settings_mfa_policy",
    "responses": {
     "200": {
      "$ref": "#/components/responses/OK"
     },
     "default": {
      "$ref": "#/components/responses/Error"
     }
    },
    "summary": "Read the Tenant's require_mfa knob (SEC-5)",
    "tags": [
     "security"
    ]
   },
   "patch": {
    "operationId": "patch_v1_admin_settings_mfa_policy",
    "responses": {
     "200": {
      "$ref": "#/components/responses/OK"
     },
     "default": {
      "$ref": "#/components/responses/Error"
     }
    },
    "summary": "Set the Tenant's require_mfa knob (tenant_admin, step-up; SEC-5)",
    "tags": [
     "security"
    ]
   }
  },
...
  "/v1/admin/users": {
   "get": {
    "operationId": "get_v1_admin_users",
    "parameters": [
     {
      "in": "query",
      "name": "role",
      "required": false,
      "schema": {
       "type": "string"
      }
     },
     {
      "in": "query",
      "name": "status",
      "required": false,
      "schema": {
       "type": "string"
      }
     },
     {
      "in": "query",
      "name": "email",
      "required": false,
      "schema": {
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "$ref": "#/components/responses/OK"
     },
     "default": {
      "$ref": "#/components/responses/Error"
     }
    },
    "summary": "List users",
    "tags": [
     "users"
    ]
   },
   "post": {
    "operationId": "post_v1_admin_users",
    "responses": {
     "201": {
      "$ref": "#/components/responses/Created"
     },
     "default": {
      "$ref": "#/components/responses/Error"
     }
    },
    "summary": "Create user (invite)",
    "tags": [
     "users"
    ]
   }
  },
  "/v1/admin/users/{id}": {
   "delete": {
    "operationId": "delete_v1_admin_users",
    "parameters": [
     {
      "in": "path",
      "name": "id",
      "required": true,
      "schema": {
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "$ref": "#/components/responses/OK"
     },
     "default": {
      "$ref": "#/components/responses/Error"
     }
    },
    "summary": "Deactivate user (soft)",
    "tags": [
     "users"
    ]
   },
   "get": {
    "operationId": "get_v1_admin_users",
    "parameters": [
     {
      "in": "path",
      "name": "id",
      "required": true,
      "schema": {
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "$ref": "#/components/responses/OK"
     },
     "default": {
      "$ref": "#/components/responses/Error"
     }
    },
    "summary": "User detail",
    "tags": [
     "users"
    ]
   },
   "patch": {
    "operationId": "patch_v1_admin_users",
    "parameters": [
     {
      "in": "path",
      "name": "id",
      "required": true,
      "schema": {
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "$ref": "#/components/responses/OK"
     },
     "default": {
      "$ref": "#/components/responses/Error"
     }
    },
    "summary": "Update role/status/ABAC attrs",
    "tags": [
     "users"
    ]
   }
  },
  "/v1/admin/users/{id}/reset-password": {
   "post": {
    "operationId": "post_v1_admin_users_reset_password",
    "parameters": [
     {
      "in": "path",
      "name": "id",
      "required": true,
      "schema": {
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "$ref": "#/components/responses/OK"
     },
     "default": {
      "$ref": "#/components/responses/Error"
     }
    },
    "summary": "Issue password reset",
    "tags": [
     "users"
    ]
   }
  },
...
 },
 "servers": [
  {
   "url": "/"
  }
 ]
}